+++ tests/backends.py
@@ -3,7 +3,7 @@ from uuid import uuid4
 
 from fastapi import Request
 from returns.maybe import Maybe
-from returns.result import Failure, Result, Success
+from returns.result import Failure, ResultE, Success
 
 from stapi_fastapi.backends.product_backend import ProductBackend
 from stapi_fastapi.backends.root_backend import RootBackend
@@ -27,15 +27,13 @@ class MockRootBackend(RootBackend):
     def __init__(self, orders: MockOrderDB) -> None:
         self._orders = orders
 
-    async def get_orders(self, request: Request) -> Result[OrderCollection, Exception]:
+    async def get_orders(self, request: Request) -> ResultE[OrderCollection]:
         """
         Show all orders.
         """
         return Success(OrderCollection(features=list(self._orders.values())))
 
-    async def get_order(
-        self, order_id: str, request: Request
-    ) -> Result[Maybe[Order], Exception]:
+    async def get_order(self, order_id: str, request: Request) -> ResultE[Maybe[Order]]:
         """
         Show details for order with `order_id`.
         """
@@ -53,7 +51,7 @@ class MockProductBackend(ProductBackend):
         product_router: ProductRouter,
         search: OpportunityRequest,
         request: Request,
-    ) -> Result[list[Opportunity], Exception]:
+    ) -> ResultE[list[Opportunity]]:
         return Success(
             [o.model_copy(update=search.model_dump()) for o in self._opportunities]
         )
@@ -63,7 +61,7 @@ class MockProductBackend(ProductBackend):
         product_router: ProductRouter,
         payload: OrderRequest,
         request: Request,
-    ) -> Result[Order, Exception]:
+    ) -> ResultE[Order]:
         """
         Create a new order.
         """